apply optimization in a different place...

+++ thesis/create_pressure.py
@@ -89,7 +89,7 @@ print(f"{x_num=}\t{y_num=}\t{t_num=}")
 
 ### Function
 def pressure(x, y, t, t0=10000., U=50., a=200000., p0=2000., x0=0.):
-    return np.float32(
+    return (
         p0
         * (1. - np.exp(-t / t0))
         * np.exp(-((x - x0)**2. + (y - U * t)**2.) / a**2.)
@@ -116,7 +116,7 @@ for case_number in range(num_cases):
     #         for k in range(x.size):
     #             p[i,j,k] = pressure(x[k], y[j], t[i], t0, U, a, p0)
 
-    p = pressure(xx, yy, tt, t0, U, a, p0, x0)
+    p = pressure(xx, yy, tt, t0, U, a, p0, x0).astype(np.float32)
 
     # remove zero-columns and zero-rows
     ix = np.where(~ np.all(np.isclose(p, 0), axis=(0, 1)))[0]
